Avoid using unsigned types to get rid of explicit casts in numeric expression.

+++ Infusion.LegacyApi/Legacy.cs
@@ -529,7 +529,7 @@ namespace Infusion.LegacyApi
             DragItem(item, item.Amount);
         }
 
-        public void DragItem(Item item, ushort amount)
+        public void DragItem(Item item, int amount)
         {
             CheckCancellation();
